api(all): cleanup JSDoc formatting and standardize API examples
The JSDoc syntax was incorrect on several matchers. This did not affect usability, but was causing
linting issues. The API documentation is now standardized, formatted well, and each matcher has an
example for testing both true and false cases. The matcher examples are also enclosed within a test
block for better demonstration of where to use it.