fix: add bounding box validity checking to geometry core functions and matchers
The spec for [bounding boxes](https://datatracker.ietf.org/doc/html/rfc7946#section-5) permits
GeoJSON objects to have bounding boxes. Section 3 of the spec states "A GeoJSON object represents a
Geometry, Feature, or collection of Features." Therefore, the geometry objects should all have
allowed bounding boxes. Thus, they all need to check that if that property is present that it is in
fact a valid 2D or 3D bbox.

Resolves: #9, #10, #11, #12, #13, #14, #16, #29