feat(toHaveStringID): add new matcher
Checks if a GeoJSON Feature has a string ID. Passes if the Feature object has any string ID (no
argument provided), or if the ID exactly matches the optional argument (single string or RegExp
provided), or any value within an array of any combination of strings or RegExp. The test fails if
the object does not have an ID, or if it has an ID that does not match the SearchID. Passing a
number type to SearchID will not pass the test, even if the ID exactly matches.

Resolves: #37