feat(toHaveNumericID): add new matcher
Checks if a GeoJSON Feature has a numeric ID. Passes if the Feature object has any numeric ID (no
argument provided), or if the ID exactly matches the optional argument (single number or RegExp
provided), or any value within an array of any combination of numbers or RegExp. The test fails if
the object does not have an ID, or if it has an ID that does not match the SearchID. Passing a
string type to SearchID will not pass the test, even if the ID exactly matches.

Resolves: #38